# pkg/trace/obfuscate: Normalize arrays and table names