--- a/releasenotes/notes/apm-group-arrays-and-quantize-table-names-276e71a2949631e2.yaml
+++ b/releasenotes/notes/apm-group-arrays-and-quantize-table-names-276e71a2949631e2.yaml
@@ -1,0 +1,14 @@
+# Each section from every release note are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+features:
+  - |
+    APM: Support SQL obfuscator feature to replace consecutive digits in table names.
+fixes:
+  - |
+    APM: Group arrays of consecutive '?' identifiers