### What does this PR do?

1. Normalize arrays and groupables made up of the `?` identifier (common from digested database queries). So that `ARRAY[?,?,?]` becomes `ARRAY [ ? ]`.
2. Normalize table names if the "normalize_sql_tables" flag is set. This is to prevent cardinality explosions like `fetch ? 
from automuted_downtimes_17_1597265206`. We may wish to make this the default behavior in the future (for instance if more than 1 digit is detected so `ec2hosts` will not be normalized but tables with timestamps will. But for now, let's put it behind a FF.


### Motivation

As a successor to: https://github.com/DataDog/datadog-agent/pull/5664, we're going to be less aggressive with the normalization we do at the agent.

These changes should have no impact on customers.

### Additional Notes

### Describe your test plan

Unit tests below