@justiniso 2020-08-25
@gbbr I did some benchmarking earlier. I can dig in more if needed, not sure what an "acceptable" threshold would be:

```
Master:

BenchmarkTokenizer/Escaping/512-8             110655       10640 ns/op      3344 B/op       93 allocs/op
BenchmarkTokenizer/Grouping/199-8             200767        6269 ns/op      2000 B/op       74 allocs/op


This branch:

BenchmarkTokenizer/Escaping/512-8             105302       11440 ns/op      3408 B/op       95 allocs/op
BenchmarkTokenizer/Grouping/199-8             195546        8242 ns/op      2048 B/op       76 allocs/op


This branch (feature enabled):

BenchmarkTokenizer/Escaping/512-8              94941       11946 ns/op      3472 B/op       97 allocs/op
BenchmarkTokenizer/Grouping/199-8             177781        7652 ns/op      2112 B/op       78 allocs/op
```