@gbbr 2020-08-27
@justiniso since obfuscation is technically hot path for some services, we need the changes on this branch to not affect performance when the feature flag isn't set. Could it perhaps be because we are now always adding the `tableFinderFilter`? I'll let you explore.

On another note, in the future it would be very helpful to post benchmark results using [`benchstat`](https://godoc.org/golang.org/x/perf/cmd/benchstat). It gives a much better perspective.