@justiniso 2020-12-15
@gbbr I'm revisiting this PR now. I believe the performance benchmarking is reporting incorrect allocs and the performance for the average case will not be impacted. Can you let me know what you think or if I'm missing something obvious?

The only difference in performance is in the number of allocs. 

```
name                               old time/op    new time/op    delta
ObfuscateSQLString/Escaping/512-8    8.88µs ± 0%    8.54µs ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/Grouping/199-8    4.70µs ± 0%    4.51µs ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/Large/3694-8       131µs ± 0%      71µs ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/random-8          1.36µs ± 0%    1.40µs ± 0%   ~     (p=1.000 n=1+1)

name                               old alloc/op   new alloc/op   delta
ObfuscateSQLString/Escaping/512-8    2.16kB ± 0%    2.17kB ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/Grouping/199-8    1.04kB ± 0%    1.05kB ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/Large/3694-8      17.0kB ± 0%    17.1kB ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/random-8            456B ± 0%      464B ± 0%   ~     (p=1.000 n=1+1)

name                               old allocs/op  new allocs/op  delta
ObfuscateSQLString/Escaping/512-8      16.0 ± 0%      27.0 ± 0%   ~     (p=1.000 n=1+1)  <---- see here
ObfuscateSQLString/Grouping/199-8      14.0 ± 0%      26.0 ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/Large/3694-8         109 ± 0%       176 ± 0%   ~     (p=1.000 n=1+1)
ObfuscateSQLString/random-8            12.0 ± 0%      13.0 ± 0%   ~     (p=1.000 n=1+1)
```

But the difference is due to code that doesn't even run during the benchmark! I created two branches to compare the line commented out (and there is a `panic` above to confirm the code doesn't execute). You can confirm the discrepancy yourself on the branches:

https://github.com/DataDog/datadog-agent/compare/justiniso/sql-normalization-fixes--case1...justiniso/sql-normalization-fixes--case2#diff-165adf1077fe63da001cd8f49ded14a19885eda516bf3b0a6fc446c65c9637f7L114-R115

#### justiniso/sql-normalization-fixes--case1
```
>>> go test -benchmem -run='^$' -bench '^(BenchmarkObfuscateSQLString)$' github.com/DataDog/datadog-agent/pkg/trace/obfuscate

BenchmarkObfuscateSQLString/Escaping/512-8 ... 23 allocs/op
BenchmarkObfuscateSQLString/Grouping/199-8 ... 22 allocs/op
BenchmarkObfuscateSQLString/Large/3694-8   ... 150 allocs/op
BenchmarkObfuscateSQLString/Complex/969-8  ... 11 allocs/op
BenchmarkObfuscateSQLString/random-8       ... 9 allocs/op
```


#### justiniso/sql-normalization-fixes--case2
```
>>> go test -benchmem -run='^$' -bench '^(BenchmarkObfuscateSQLString)$' github.com/DataDog/datadog-agent/pkg/trace/obfuscate

BenchmarkObfuscateSQLString/Escaping/512-8 ... 12 allocs/op
BenchmarkObfuscateSQLString/Grouping/199-8 ... 10 allocs/op
BenchmarkObfuscateSQLString/Large/3694-8   ... 83 allocs/op
BenchmarkObfuscateSQLString/Complex/969-8  ... 6 allocs/op
BenchmarkObfuscateSQLString/random-8       ... 8 allocs/op
```